+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#373](https://github.com/spegel-org/spegel/pull/373) Apply mirror configuration on all registires by default.
 - [#762](https://github.com/spegel-org/spegel/pull/762) Set appropriate buckets for response size
 - [#778](https://github.com/spegel-org/spegel/pull/778) Replace interface{} with any alias.
+- [#784](https://github.com/spegel-org/spegel/pull/784) Refactor distribution and move to OCI package.
 
 ### Deprecated